+++ src/plugins/core/request/index.ts
@@ -1,5 +1,6 @@
 export * from './attachDefaultHeaders';
 export * from './encodeRequestBody';
+export * from './excludeProxyInterceptor';
 export * from './isTrustedInterceptor';
 export * from './requestVariableReplacer';
 export * from './setEnvRequestOptions';